Fix soma table rule message & Move merge_operation.py
See #26
