+++ nuclei_prediction/README.md
@@ -118,4 +118,4 @@ Even though we manually checked our somas' quality multiple times, there is a sm
 At the same time, our convolutional neural network missed some nuclei. (e.g., A neuronal nucleus that was partically detected but too small to be picked up by our size threshold.) This is mostly because of the knifemarks, and DeepEM could not segment a nucleus correctly if there is a knifemark on it. Therefore, we asked community members to report if they have found any nuclei that were not listed on the soma table. Those manually identified nuclei have their own 17-digit nucleus ids, which starts from `10000000000000001`. If you are a member of `FANC_soma_edit` group on CAVE, you can use the code in `fanc/upload.py` to upload these missing somas to the soma table. `example_notebooks/update_cave_tables.ipynb` demostrates how to upload them. Currently, @sumiya-kuroda and @jasper-tms are the members of `FANC_soma_edit`.
 
 There are also some other rules when it comes to uploading manually detected somas:
-- If you want to upload neurons that have somas outside the dataset and locate on the ventral side, use coordinates on `z = 10` slice.
+- If you want to upload neurons that have somas outside the dataset and locate on the dorsal side, use coordinates on `z = 10` slice.
