+++ nuclei_prediction/format/merge_operation.py
